feat(portal-frontend): klai/no-cross-route-import ESLint rule + extract KBOverviewSections
Two intertwined changes — the rule needs the extraction to be enforceable
without false positives.

1. New ESLint rule `klai/no-cross-route-import` (with 13-case unit test)
   that enforces the "cross-route imports are a smell" anti-pattern
   documented in portal-frontend.md § "File organization for shared
   types and helpers". Heuristic:
   - importing file lives under src/routes/ AND its basename / parent
     dirs do NOT contain colocation markers (`-` prefix, `_` prefix,
     `._` infix)
   - import source is relative AND none of its segments contain
     colocation markers
   → REPORT.
   Allowed targets: `-`-prefixed siblings/files, `_components/` dirs,
   `<route>._<feature>` colocation, `routeTree.gen`, `__tests__/`.

2. KBOverviewSections extracted from `overview.tsx` (route file, was
   1 lint violation source) to `_components/KBOverviewSections.tsx`
   (TanStack-ignored directory). insights.tsx now imports from the
   `_components/` location → no rule violation.

3. TaxonomyTab cross-route import in insights.tsx is marked with
   `eslint-disable-next-line` + a TODO referencing
   SPEC-PORTAL-CONNECTOR-WIZARD-EXTRACT-001 § Follow-ups F-table
   row 1. Splitting the 720-line god-component is its own SPEC and
   not the right scope for this cleanup branch. The disable comment
   is the explicit deferred-fix marker — when F-table row 1 lands,
   the disable goes away.

The connector wizard refactor itself contains zero cross-route imports,
verified by the rule running clean against all wizard files. The rule
catches future regressions automatically.

Verification:
- vitest run (full suite): 235/235 passing
- new rule's tests: 13/13 passing (4 valid + 4 invalid + 5 edge cases)
- tsc --noEmit clean (only pre-existing tsconfig warning)
- eslint . clean (exit 0, 0 messages)
